tests/wpt/mozilla/tests/accessibility-tree/accessibility-update-partial-subtree-move-and-delete.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<section>
  <div id="div1">
    <div id="div2">
      <h1 id="h1">This is an h1</h1>
      <p id="p">This is a paragraph</p>
    </div>
  </div>
  <div id="div3"><h2 id="h2">This is an h2</h2></div>
</section>

<script>
  setup({ explicit_done: true });
  window.onload = () => {
    test((t) => {
      ServoTestUtils.ensureAccessibilityActive();
      let initial = ServoTestUtils.forceAccessibilityUpdate();

      assert_equals(
        initial.accessibilityNodesUpdatedFromDom,
        29,
        "accessibilityNodesUpdatedFromDom",
      );
      assert_equals(
        initial.accessibilityNodesUpdatedFromTree,
        29,
        "accessibilityNodesUpdatedFromTree",
      );
      assert_equals(
        initial.accessibilityNodesInTreeUpdate,
        29,
        "accessibilityNodesInTreeUpdate",
      );

      div3.moveBefore(div2, h2);
      div3.moveBefore(h1, div2);
      p.remove();
      div1.remove();

      let incremental = ServoTestUtils.forceAccessibilityUpdate();

      // Updated nodes:
      // - div1 (div2 removed - redundant update, since div1 is removed!)
      // - <section> (div1 removed)
      // - div2 (<p> and <h1> removed)
      // - div3 (new children)
      assert_equals(
        incremental.accessibilityNodesUpdatedFromDom,
        4,
        "accessibilityNodesUpdatedFromDom",
      );

      // <html>, <body>, <section>, div3, div2
      assert_equals(
        incremental.accessibilityNodesUpdatedFromTree,
        5,
        "accessibilityNodesUpdatedFromTree",
      );

      // Changed nodes:
      // - <section>, children changed
      // - div2, children changed
      // - div3, children changed
      assert_equals(
        incremental.accessibilityNodesInTreeUpdate,
        3,
        "accessibilityNodesInTreeUpdate",
      );
    }, "");

    done();
  };
</script>
